Minor refactor to improve the UI and make the banking utilities more flexible
