+++ AIO/src/org/aio/activities/skills/thieving/ThievingActivity.java
@@ -139,7 +139,7 @@ public class ThievingActivity extends Activity {
 
     private class ThievingBank extends Banking {
         @Override
-        public void bank() {
+        public boolean bank() {
             if (!getInventory().isEmpty()) {
                 getBank().depositAll();
             } else if (food != Food.NONE) {
@@ -149,6 +149,8 @@ public class ThievingActivity extends Activity {
                     setFailed();
                 }
             }
+
+            return true;
         }
     }
 }
