+++ AIO/src/org/aio/gui/task_panels/BreakTaskPanel.java
@@ -2,8 +2,8 @@ package org.aio.gui.task_panels;
 
 import org.aio.gui.utils.NumberDocumentFilter;
 import org.aio.tasks.break_task.BreakTask;
-import org.aio.tasks.task.Task;
-import org.aio.tasks.task.TaskType;
+import org.aio.tasks.Task;
+import org.aio.tasks.TaskType;
 import org.json.simple.JSONObject;
 
 import javax.swing.*;
