+++ AIO/src/org/aio/gui/task_panels/GETaskPanel.java
@@ -3,9 +3,9 @@ package org.aio.gui.task_panels;
 import org.aio.activities.grand_exchange.*;
 import org.aio.gui.utils.AutoCompleteTextField;
 import org.aio.gui.utils.NumberDocumentFilter;
-import org.aio.tasks.grand_exchange_task.GrandExchangeTask;
-import org.aio.tasks.task.Task;
-import org.aio.tasks.task.TaskType;
+import org.aio.tasks.GrandExchangeTask;
+import org.aio.tasks.Task;
+import org.aio.tasks.TaskType;
 import org.json.simple.JSONObject;
 
 import javax.swing.*;
