+++ AIO/src/org/aio/gui/task_panels/TaskPanel.java
@@ -1,7 +1,7 @@
 package org.aio.gui.task_panels;
 
 import org.aio.gui.JSONSerializable;
-import org.aio.tasks.task.Task;
+import org.aio.tasks.Task;
 
 import javax.swing.*;
