+++ AIO/src/org/aio/script/AIO.java
@@ -5,8 +5,8 @@ import org.aio.gui.conf_man.ConfigManager;
 import org.aio.paint.MouseTrail;
 import org.aio.paint.Paint;
 import org.aio.tasks.break_task.CustomBreakManager;
-import org.aio.tasks.task.Task;
-import org.aio.tasks.task.TaskType;
+import org.aio.tasks.Task;
+import org.aio.tasks.TaskType;
 import org.aio.util.SkillTracker;
 import org.aio.util.event.ToggleShiftDropEvent;
 import org.json.simple.JSONObject;
