+++ AIO/src/org/aio/tasks/LevelTask.java
@@ -1,8 +1,6 @@
-package org.aio.tasks.level_task;
+package org.aio.tasks;
 
 import org.aio.activities.activity.Activity;
-import org.aio.tasks.task.Task;
-import org.aio.tasks.task.TaskType;
 import org.osbot.rs07.api.ui.Skill;
 
 public class LevelTask extends Task {
