+++ AIO/src/org/aio/tasks/TutorialIslandTask.java
@@ -1,8 +1,8 @@
-package org.aio.tasks.tutorial_island_task;
+package org.aio.tasks;
 
 import org.aio.activities.tutorial_island.TutorialIsland;
-import org.aio.tasks.task.Task;
-import org.aio.tasks.task.TaskType;
+import org.aio.tasks.Task;
+import org.aio.tasks.TaskType;
 
 public class TutorialIslandTask extends Task {
