+++ AIO/src/org/aio/tasks/task_executor/TaskChangeListener.java
@@ -1,6 +1,6 @@
 package org.aio.tasks.task_executor;
 
-import org.aio.tasks.task.Task;
+import org.aio.tasks.Task;
 
 public interface TaskChangeListener {
     void taskChanged(final Task oldTask, final Task newTask);
